requirements.txt
PyGLM==1.99.2
PyOpenGL==3.1.5
PyOpenGL-accelerate==3.1.5
PySide2==5.15.0
shiboken2==5.15.0
